% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buy_check.R
\name{buy_check}
\alias{buy_check}
\title{Buy check}
\usage{
buy_check(granuleId)
}
\arguments{
\item{granuleId}{a character vector of one or more granuleId's}
}
\value{
logical of the same length as granuleId, where \code{TRUE} means a
  granule needs to be bougth to gain access
}
\description{
Checks, if granule's are already available or need to be bought
}

