% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date.R
\name{check_date}
\alias{check_date}
\title{Check user supplied date string}
\usage{
check_date(x)
}
\arguments{
\item{x}{a character string expected to have format 'YYYY-MM-DD'}
}
\value{
object of class 'Date', if x has a valid date of format 'YYYY-MM-DD',
  otherwise an error is thrown
}
\description{
Will check if date string is well formated and throw and error if not
}

