% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{.onLoad setup credentials}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{library name}

\item{pkgname}{package name}
}
\description{
Setup login credentials on package start, if not retrieved already (e.g. from
 .Rprofile).
}
