% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date.R
\name{format_date}
\alias{format_date}
\title{Reformats date}
\usage{
format_date(x)
}
\arguments{
\item{x}{a character string representing date/time as 'YYYY-MM-DD HH:MM:SS'}
}
\value{
character of format 'YYYY-MM-DD'
}
\description{
Simplifies the 'date' from 'YYYY-MM-DD HH:MM:SS' to 'YYYY-MM-DD' for
filenaming
}

