% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_to_geometry.R
\name{geojson_to_geometry}
\alias{geojson_to_geometry}
\title{Converts a geoJSON string into spatial objects used by R packages}
\usage{
geojson_to_geometry(geojson, package)
}
\arguments{
\item{geojson}{character vector of geoJSON objects}

\item{package}{name of an R package to to the format used by which
\code{geojson} geometries should be coverted. Currently \code{sp} and
\code{sf} packages are supported}
}
\value{
a list of spatial objects
}
\description{
Converts a geoJSON string into spatial objects used by R packages
}
