% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgeom_to_SpatialPolygons.R
\name{jgeom_to_SpatialPolygons}
\alias{jgeom_to_SpatialPolygons}
\title{Converts json geometry to SpatialPolygons}
\usage{
jgeom_to_SpatialPolygons(x)
}
\arguments{
\item{x}{character json geometry as returned by a database query with
'retGeometry = 1'}
}
\value{
SpatialPolygons object of the geometry supplied via x
}
\description{
Turn a json geometry as returned by a database query using 'retGeometry = 1'
  into a SpatialPolygons object
}
\note{
The geometry is first replaced by its convex hull for simplification!
}

