% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi2jgeom.R
\name{roi2jgeom}
\alias{roi2jgeom}
\title{Convert a region of interest to a json geometry}
\usage{
roi2jgeom(roi, projection = sp::CRS("+init=epsg:4326"))
}
\arguments{
\item{roi}{a region of interest supplied as a vector/matrix of named ('x',
'y') coordinates or any sp package spatial
object or a path to a file \code{\link[rgdal]{readOGR}} can open or a json
geometry character string}

\item{projection}{sp's package CRS object providing points projection.
Applicable only when the \code{roi} parameter is a vector/matrix of
coordinates}
}
\value{
character, a json geometry string
}
\description{
Convert a region of interest to a json geometry
}
\note{
Coordinates are assumed to be projected (WGS-84, 4326). Spatial objects
  will be reprojected as necessary
}
