% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_to_jgeom.R
\name{roi_to_jgeom}
\alias{roi_to_jgeom}
\title{Convert a region of interest to a json geometry}
\usage{
roi_to_jgeom(roi, round = Inf)
}
\arguments{
\item{roi}{a region of interest supplied as a vector/matrix of named ('x', 'y')
coordinates or a SpatialPoints object or a SpatialPolygons object or a path
to a file \code{\link[rgdal]{readOGR}} can open or a json geometry character string}

\item{round}{number of digits to round coordinates to}
}
\value{
character, a json geometry string
}
\description{
Convert a region of interest to a json geometry
}
\note{
Coordinates are assumed to be projected (WGS-84, 4326). Spatial objects
  will be reprojected as necessary
}
