% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_buy_granule.R
\name{S2_buy_granule}
\alias{S2_buy_granule}
\title{Buy a granule at 'https://s2.boku.eodc.eu'}
\usage{
S2_buy_granule(granuleId)
}
\arguments{
\item{granuleId}{character the granuleId}
}
\description{
Buy granule using granuleId
}

