% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_check_access.R
\name{S2_check_access}
\alias{S2_check_access}
\title{(Very basic) check if server/database can be accessed}
\usage{
S2_check_access(verbose = TRUE)
}
\arguments{
\item{verbose}{logical}
}
\value{
logical \code{TRUE} if access to server was successfull and user
  credentials seem to be valid, FALSE if 'anything' went wrong
}
\description{
Checks accessibility of server / database and does basic checking of
  user credentials
}

