% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_do_query.R
\name{S2_do_query}
\alias{S2_do_query}
\title{Send a query to 'https://s2.boku.eodc.eu'}
\usage{
S2_do_query(query, path, baseUrl = "https://s2.boku.eodc.eu")
}
\arguments{
\item{query}{list of named arguments 'known' to the database. Typically
created using e.g. 'S2_query_granule'}

\item{path}{name of the rest API endpoint like "product", "granule", "image",
"qiData", "angle", "roi", "job" or "rgb".}

\item{baseUrl}{base REST API URL}
}
\description{
Query database and return response as data.frame
}
