% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_download.R
\name{S2_download}
\alias{S2_download}
\title{S2_download helper function}
\usage{
S2_download(url, destfile, zip = TRUE)
}
\arguments{
\item{url}{character (valid) url to download file from.}

\item{destfile}{character download destination. If \code{zip = TRUE}, the
ending '.zip' will be attached to destfile (, if it is not already).}

\item{zip}{logical if \code{TRUE}, the url will be downloaded as zip archive
and (automatically) unzipped in the parent directory of 'destfile'}
}
\description{
Simple helper function to zip download and unzip i.e. a 'granule' in a single
  step
}

