% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_generate_names.R
\name{S2_generate_names}
\alias{S2_generate_names}
\title{S2_granule_naming}
\usage{
S2_generate_names(x, ...)
}
\arguments{
\item{x}{is the returned data \code{S2_query_granule} or \code{S2_query_image}}

\item{...}{arguments passed to S2_naming()}
}
\value{
character names
}
\description{
Creates a filename to store a downloaded 'granule', based on query response
for a single 'granuleId'
}
\details{
This function allows to alter the automatic naming behavior for
  downloaded granules. Useful, when it is desired to add a custom prefix /
  suffix or make files be easier sortable by utm zone Id instead of
  acquision date.
}

