% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_initialize_user.R
\name{S2_initialize_user}
\alias{S2_initialize_user}
\title{Function used to set 'user' and 'password' login credentials for the current
  R session}
\usage{
S2_initialize_user(user = "default@foo.bar", password = "default")
}
\arguments{
\item{user}{character user login you have registered with at
'https://s2.boku.eodc.eu'}

\item{password}{character password for 'https://s2.boku.eodc.eu'}
}
\description{
S2_initialize_user stores 'user' and 'password' usign \code{options()}.
 Functions in package 'S2boku' (that require authentication) will retrieve
 the credentials using 'getOption()'. This will avoid passing login
 credentials multiple times in a single session and allows the user
 for permanent configuration in '.Rprofile'.
}
\note{
Requires a valid registration to 'https://s2.boku.eodc.eu' to gain
   access to database functionality where authentication is mandatory.
}

