% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_query_granule.R
\name{S2_query_granule}
\alias{S2_query_granule}
\title{Prepare a list to query database for 'granule'}
\usage{
S2_query_granule(atmCorr = NULL, broken = 0, cloudCovMin = 0,
  cloudCovMax = 100, dateMax = Sys.Date(), dateMin = as.Date(dateMax) -
  365, geometry = NULL, granule = NULL, granuleId = NULL,
  orbitNo = NULL, product = NULL, productId = NULL, retGeometry = 0,
  utm = NULL, dateSingle = NULL, ...)
}
\arguments{
\item{atmCorr}{see wiki}

\item{broken}{see wiki}

\item{cloudCovMin}{see wiki}

\item{cloudCovMax}{see wiki}

\item{dateMax}{see wiki}

\item{dateMin}{see wiki}

\item{geometry}{see wiki}

\item{granule}{see wiki}

\item{granuleId}{see wiki}

\item{orbitNo}{see wiki}

\item{product}{see wiki}

\item{productId}{see wiki}

\item{retGeometry}{see wiki}

\item{utm}{see wiki}

\item{dateSingle}{see wiki}

\item{...}{further arguments, none implemented}
}
\value{
list of query arguments
}
\description{
Implements the query options for granule found in the wiki
}

