% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_query_image.R
\name{S2_query_image}
\alias{S2_query_image}
\title{Prepare a list to query database for 'image'}
\usage{
S2_query_image(atmCorr = NULL, band = NULL, broken = 0, cloudCovMin = 0,
  cloudCovMax = 100, dateMax = Sys.Date(), dateMin = as.Date(dateMax) -
  365, geometry = NULL, granule = NULL, granuleId = NULL,
  imageId = NULL, orbitNo = NULL, product = NULL, productId = NULL,
  resolution = NULL, retGeometry = 0, utm = NULL, dateSingle = NULL,
  ...)
}
\arguments{
\item{atmCorr}{see wiki}

\item{band}{see wiki}

\item{broken}{see wiki}

\item{cloudCovMin}{see wiki}

\item{cloudCovMax}{see wiki}

\item{dateMax}{see wiki}

\item{dateMin}{see wiki}

\item{geometry}{see wiki}

\item{granule}{see wiki}

\item{granuleId}{see wiki}

\item{imageId}{see wiki}

\item{orbitNo}{see wiki}

\item{product}{see wiki}

\item{productId}{see wiki}

\item{resolution}{see wiki}

\item{retGeometry}{see wiki}

\item{utm}{see wiki}

\item{dateSingle}{see wiki}

\item{...}{further arguments, none implemented}
}
\value{
list of query arguments
}
\description{
Implements the query options for 'images' found in the wiki
}

