% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_query_job.R
\name{S2_query_job}
\alias{S2_query_job}
\title{Prepare a list to query database for 'job'}
\usage{
S2_query_job(dateMax = Sys.Date(), dateMin = as.Date(dateMax) - 365,
  ended = 1, failed = 0, geometry = NULL, jobId = NULL,
  granule = NULL, granuleId = NULL, product = NULL, productId = NULL,
  retGeometry = 0, started = NULL, dateSingle = NULL, ...)
}
\arguments{
\item{dateMax}{see wiki}

\item{dateMin}{see wiki}

\item{ended}{see wiki}

\item{failed}{see wiki}

\item{geometry}{see wiki}

\item{jobId}{see wiki}

\item{granule}{see wiki}

\item{granuleId}{see wiki}

\item{product}{see wiki}

\item{productId}{see wiki}

\item{retGeometry}{see wiki}

\item{started}{see wiki}

\item{dateSingle}{see wiki}

\item{...}{further arguments, none implemented}
}
\value{
list of query arguments
}
\description{
Implements the query options for 'job' found in the wiki
}

