% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_user_info.R
\name{S2_user_info}
\alias{S2_user_info}
\title{S2_user_info}
\usage{
S2_user_info()
}
\value{
list where 'userId' is the Id of the current user, 'coins' is the
  total number of coins acquired by the user, 'coinsUsed' the number of coins
  used so far and 'coinsRemain' the number of coins available to acquire
  database products
}
\description{
Returns information on a users coins budget
}
