% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_vrt_L2Agranule.R
\name{S2_vrt_L2Agranule}
\alias{S2_vrt_L2Agranule}
\title{Create a virtual raster from a (Level 2A) Sentinel-2 Granule}
\usage{
S2_vrt_L2Agranule(granule, verbose = FALSE)
}
\arguments{
\item{granule}{character path to a 'granule'}

\item{verbose}{logical if \code{TRUE} show warnings}
}
\value{
path to the created file
}
\description{
Creates a '.vrt' raster from a granule directory
}
\note{
Requires a valid 'GDAL' install (and properly configured '$PATH' variable).
  Suppose a granule is complete (containing L2A files at 10m, 20m and 60m resolution),
  the resulting '.vrt' stack will contain bands in the following order:
  B01 (60m), B02 (10m), B03 (10m), B04 (10m), B05 (20m), B06 (20m), B07 (20m),
  B08 (10m), B09 (60m), B11 (20m) and B12 (20m).
  For details, please consult the ESA Sentinel-2 product specifications documentation.
}
