% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat2jgeom.R
\name{spat2jgeom}
\alias{spat2jgeom}
\title{Convert any sp package geometry to a json geometry}
\usage{
spat2jgeom(spat)
}
\arguments{
\item{spat}{an sp package spatial object or list of sp package spatial
objects}
}
\value{
vector of corresponding JSON geometries
}
\description{
Convert any sp package geometry to a json geometry
}
\note{
Objects consisting of many one feature are covnerted to GeoJSON
Polygons while objects consisting of many features to MultiPolygons.
}
