% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_do_query.R
\name{S2_do_query}
\alias{S2_do_query}
\title{Send a query to 'https://s2.boku.eodc.eu'}
\usage{
S2_do_query(query, path)
}
\arguments{
\item{query}{list of named arguments 'known' to the database. Typically
created using e.g. 'S2_query_granule'}

\item{path}{character altering the query url with respect to the desired
output. One of "product", "granule", "image", "qiData", "angle", "roi",
"job", "rgb".}
}
\description{
Query database and return response as data.frame
}
