% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_download.R
\name{S2_download}
\alias{S2_download}
\title{S2_download helper function}
\usage{
S2_download(url, destfile, zip = TRUE, skipExisting = TRUE, ...)
}
\arguments{
\item{url}{character (valid) url to download file from.}

\item{destfile}{character download destination.}

\item{zip}{logical if \code{TRUE}, the url will be downloaded as zip archive
and (automatically) unzipped in the parent directory of 'destfile'
(plays any role only when downloading granules).}

\item{skipExisting}{logical skip if file already exists.}

\item{...}{further arguments not implemented directly - see
the \href{https://s2.boku.eodc.eu/wiki/#!granule.md#GET_https://s2.boku.eodc.eu/granule/{granuleId}}{granule API doc}
and the \href{https://s2.boku.eodc.eu/wiki/#!image.md#GET_https://s2.boku.eodc.eu/image/{imageId}}{image API doc}.}
}
\value{
logical vector indicating which downloads where successful
}
\description{
Simple helper function to download data (granules, images, QI data)
}
\examples{
\dontrun{
  # find, download and unzip a full granule
  granules <- S2_query_granule(
    utm = '33UXP',
    dateMin = '2016-06-01',
    dateMax = '2016-06-30'
  )
  S2_download(granules$url, granules$date)

  # find and download a bunch of images
  images <- S2_query_image(
    utm = '33UXP',
    dateMin = '2016-06-01',
    dateMax = '2016-06-30',
    band = 'B03'
  )
  S2_download(images$url, paste0(images$date, '.', images$format))

  # reproject downloaded images by passing additional API parameter
  S2_download(images$url, paste0(images$date, '.', images$format), srid = 4326)

  # download particular URL
  S2_download(
    'https://test\%40s2\%2Eboku\%2Eeodc\%2Eeu:test@s2.boku.eodc.eu/image/33148479',
    'test.jp2'
  )
}
}
