% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_initialize_user.R
\name{S2_initialize_user}
\alias{S2_initialize_user}
\title{Set login credentials for 's2.boku.eodc.eu'}
\usage{
S2_initialize_user(user = "test@s2.boku.eodc.eu", password = "test")
}
\arguments{
\item{user}{character user login you have registered with at
'https://s2.boku.eodc.eu'}

\item{password}{character password for 'https://s2.boku.eodc.eu'}
}
\description{
Set 'user' and 'password' login credentials for the current
  R session
}
\details{
\code{S2_initialize_user} store 'user' and 'password' in \code{options}.
 If needed, functions in package 'sentinel2' will retrieve the credentials
 using 'getOption()'. This will avoid passing login credentials multiple
 times in a single session and allows users to permanentely configure their
 login in a '.Rprofile' file.
}
\note{
Requires a valid registration to 'https://s2.boku.eodc.eu' to gain
   access to database functionality where authentication is mandatory.
}
