% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_query_granule.R
\name{S2_query_granule}
\alias{S2_query_granule}
\title{Query database for 'granule'}
\usage{
S2_query_granule(atmCorr = NULL, broken = FALSE, cloudCovMin = 0,
  cloudCovMax = 100, dateMax = Sys.Date(), dateMin = as.Date(dateMax) -
  365, geometry = NULL, granule = NULL, granuleId = NULL,
  orbitNo = NULL, owned = FALSE, product = NULL, productId = NULL,
  retGeometry = FALSE, utm = NULL, dateSingle = NULL, spatial = FALSE,
  ...)
}
\arguments{
\item{atmCorr}{logical if TRUE only results for atmospherically corrected
data are returned.}

\item{broken}{logical, TRUE if the granule is marked as broken
(almost for sure you want to use FALSE, the default).}

\item{cloudCovMin}{integer minimum cloud coverage from 0 to 100.}

\item{cloudCovMax}{integer maximum cloud coverage from 0 to 100.}

\item{dateMax}{character end date of format "YYYY-MM-DD".}

\item{dateMin}{character start date of format "YYYY-MM-DD".}

\item{geometry}{geometry which should intersect with granules. Can be a
geoJSON geometry string (e.g. {"type":"Point","coordinates":[16.5,48.5]}),
the path to a Point/Polygon shapefile, a SpatialPoints object or a
SpatialPolygons object.}

\item{granule}{character ESA granule id.}

\item{granuleId}{integer internal metadata database granule id.}

\item{orbitNo}{integer from 1 to 143.}

\item{owned}{logical when TRUE only already bought granules will be returned.}

\item{product}{chracter ESA product id.}

\item{productId}{integer internal metadata database product id.}

\item{retGeometry}{logical should product geometry be included in the response?}

\item{utm}{character UTM zone, e.g. 33U, 01C.}

\item{dateSingle}{character date of format "YYYY-MM-DD", specifies a single
date and will override \code{dateMin} and \code{dateMax}.}

\item{spatial}{logical, if TRUE an object of class SpatialPolygonsDataFrame
is returned.}

\item{...}{further arguments, none implemented.}
}
\value{
data.frame return of the database. If \code{spatial = TRUE} an object
  of class SpatialPolygonsDataFrame is returned.
}
\description{
Implements the query options for granule described in the wiki.
}

