% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_query_image.R
\name{S2_query_image}
\alias{S2_query_image}
\title{Prepare a list to query database for 'image'}
\usage{
S2_query_image(atmCorr = NULL, band = NULL, broken = FALSE,
  cloudCovMin = NULL, cloudCovMax = NULL, dateMax = NULL,
  dateMin = NULL, geometry = NULL, granule = NULL,
  granuleId = NULL, imageId = NULL, orbitNo = NULL, owned = FALSE,
  product = NULL, productId = NULL, regionId = NULL,
  resolution = NULL, retGeometry = FALSE, utm = NULL,
  dateSingle = NULL, spatial = FALSE, ...)
}
\arguments{
\item{atmCorr}{logical if TRUE only results for atmospherically corrected
data are returned.}

\item{band}{chracter spectral band: B01/B02/B03/B04/B05/B06/B07/B08/B8A/B09/
B10/B11/B12/all/SCL/LAI/albedo - see sentinel2 documentation.}

\item{broken}{logical, TRUE if the granule is marked as broken
(almost for sure you want to use FALSE, the default).}

\item{cloudCovMin}{integer minimum cloud coverage from 0 to 100.}

\item{cloudCovMax}{integer maximum cloud coverage from 0 to 100.}

\item{dateMax}{character end date of format "YYYY-MM-DD".}

\item{dateMin}{character start date of format "YYYY-MM-DD".}

\item{geometry}{geometry which should intersect with granules. Can be a
geoJSON geometry string (e.g. {"type":"Point","coordinates":[16.5,48.5]}),
the path to a Point/Polygon shapefile, a SpatialPoints object or a
SpatialPolygons object.}

\item{granule}{character ESA granule id.}

\item{granuleId}{integer internal metadata database granule id.}

\item{imageId}{integer internal metadata database image id.}

\item{orbitNo}{integer from 1 to 143.}

\item{owned}{logical when TRUE only already bought granules will be returned.}

\item{product}{chracter ESA product id.}

\item{productId}{integer internal metadata database product id.}

\item{regionId}{region of interest id (overrides the \code{geometry} parameter,
if \code{dateMin}, \code{dateMax} or \code{cloudCovMax} are not specified,
they are taken from the region of interest settings)}

\item{resolution}{integer spatial resolution [m]: 60/20/10
(resolution depends on the spectral band).}

\item{retGeometry}{logical should product geometry be included in the response?}

\item{utm}{character UTM zone, e.g. 33U, 01C.}

\item{dateSingle}{character date of format "YYYY-MM-DD", specifies a single
date and will override \code{dateMin} and \code{dateMax}.}

\item{spatial}{logical, if TRUE the geometry column will contain
SpatialPolygonsDataFrame objects instead of GeoJSON strings.
Be aware that such conversion may take quite some time for large number of
returned granules.}

\item{...}{further arguments not implemented directly - see
the \href{https://s2.boku.eodc.eu/wiki/#!image.md#GET_https://s2.boku.eodc.eu/image}{API doc}.}
}
\value{
data.frame return of the database.
}
\description{
Implements the query options for 'images' described in the wiki.
}
