% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_query_image.R
\name{S2_query_image}
\alias{S2_query_image}
\title{Prepare a list to query database for 'image'}
\usage{
S2_query_image(atmCorr = NULL, band = NULL, broken = FALSE,
  cloudCovMin = 0, cloudCovMax = 100, dateMax = Sys.Date(),
  dateMin = as.Date(dateMax) - 365, geometry = NULL, granule = NULL,
  granuleId = NULL, imageId = NULL, orbitNo = NULL, product = NULL,
  productId = NULL, resolution = NULL, retGeometry = FALSE, utm = NULL,
  dateSingle = NULL, ...)
}
\arguments{
\item{atmCorr}{logical if TRUE only results for atmospherically corrected
data are returned.}

\item{band}{chracter spectral band: B01/B02/B03/B04/B05/B06/B07/B08/B8A/B09/
B10/B11/B12/all/SCL/LAI/albedo - see sentinel2 documentation.}

\item{broken}{logical, TRUE if the granule is marked as broken
(almost for sure you want to use FALSE, the default).}

\item{cloudCovMin}{integer minimum cloud coverage from 0 to 100.}

\item{cloudCovMax}{integer maximum cloud coverage from 0 to 100.}

\item{dateMax}{character end date of format "YYYY-MM-DD".}

\item{dateMin}{character start date of format "YYYY-MM-DD".}

\item{geometry}{geometry which should intersect with granules. Can be a
geoJSON geometry string (e.g. {"type":"Point","coordinates":[16.5,48.5]}),
the path to a Point/Polygon shapefile, a SpatialPoints object or a
SpatialPolygons object.}

\item{granule}{character ESA granule id.}

\item{granuleId}{integer internal metadata database granule id.}

\item{imageId}{integer internal metadata database image id.}

\item{orbitNo}{integer from 1 to 143.}

\item{product}{chracter ESA product id.}

\item{productId}{integer internal metadata database product id.}

\item{resolution}{integer spatial resolution [m]: 60/20/10
(resolution depends on the spectral band).}

\item{retGeometry}{logical should product geometry be included in the response?}

\item{utm}{character UTM zone, e.g. 33U, 01C.}

\item{dateSingle}{character date of format "YYYY-MM-DD", specifies a single
date and will override \code{dateMin} and \code{dateMax}.}

\item{...}{further arguments, none implemented.}
}
\value{
data.frame return of the database.
}
\description{
Implements the query options for 'images' described in the wiki.
}

