% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_query_job.R
\name{S2_query_job}
\alias{S2_query_job}
\title{Query database for 'job'}
\usage{
S2_query_job(dateMax = Sys.Date(), dateMin = "2000-01-01", started = NULL,
  ended = NULL, failed = NULL, geometry = NULL, jobId = NULL,
  granule = NULL, granuleId = NULL, product = NULL, productId = NULL,
  retGeometry = FALSE, dateSingle = NULL, ...)
}
\arguments{
\item{dateMax}{character end date of format "YYYY-MM-DD".}

\item{dateMin}{character start date of format "YYYY-MM-DD".}

\item{started}{logical, TRUE if job has started.}

\item{ended}{logical, TRUE if job has ended.}

\item{failed}{logical, TRUE if job has failed.}

\item{geometry}{geometry which should intersect with granules. Can be a
geoJSON geometry string (e.g. {"type":"Point","coordinates":[16.5,48.5]}),
the path to a Point/Polygon shapefile, a SpatialPoints object or a
SpatialPolygons object.}

\item{jobId}{character internal metadata database job id.}

\item{granule}{character ESA granule id.}

\item{granuleId}{integer internal metadata database granule id.}

\item{product}{chracter ESA product id.}

\item{productId}{integer internal metadata database product id.}

\item{retGeometry}{logical should product geometry be included in the response?}

\item{dateSingle}{character date of format "YYYY-MM-DD", specifies a single
date and will override \code{dateMin} and \code{dateMax}.}

\item{...}{further arguments, none implemented.}
}
\value{
data.frame return of the database.
}
\description{
Implements the query options for 'job' described in the wiki.
}
