% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2_query_product.R
\name{S2_query_product}
\alias{S2_query_product}
\title{Prepare a list to query database for 'product'}
\usage{
S2_query_product(atmCorr = NULL, dateMax = Sys.Date(),
  dateMin = as.Date(dateMax) - 365, geometry = NULL, orbitDir = NULL,
  orbitNo = NULL, product = NULL, productId = NULL, retGeometry = 0,
  dateSingle = NULL, ...)
}
\arguments{
\item{atmCorr}{see wiki}

\item{dateMax}{see wiki}

\item{dateMin}{see wiki}

\item{geometry}{see wiki}

\item{orbitDir}{see wiki}

\item{orbitNo}{see wiki}

\item{product}{see wiki}

\item{productId}{see wiki}

\item{retGeometry}{see wiki}

\item{dateSingle}{see wiki}

\item{...}{further arguments, none implemented}
}
\value{
list of query arguments
}
\description{
Implements the query options for 'product' found in the wiki
}

