% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_to_jgeom.R
\name{spat_to_jgeom}
\alias{spat_to_jgeom}
\title{Convert SpatialPoints/PatialPolygons to json geometry}
\usage{
spat_to_jgeom(spat, round = Inf)
}
\arguments{
\item{spat}{a spatial object of class SpatialPolygons or SpatialPoints}

\item{round}{number of digits to round coordinates to}
}
\value{
character string, a single json geometry
}
\description{
Turn a SpatialPoints/PatialPolygons object into a json geometry string
}

